refactor path chat
