+++ src/components/Common/SpeedDial/index.js
@@ -7,7 +7,7 @@ import HomeIcon from '@material-ui/icons/Home';
 import SpeedDial from '@material-ui/lab/SpeedDial';
 import SpeedDialAction from '@material-ui/lab/SpeedDialAction';
 import SpeedDialIcon from '@material-ui/lab/SpeedDialIcon';
-import { LINKS } from '../../constant';
+import { LINKS } from '../../../constant';
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 // import { Link } from 'react-scroll';
