+++ src/components/index.js
@@ -2,7 +2,7 @@ export { default as Drawer } from "./Class/Drawer/Drawer";
 export { default as Header } from "./Class/Header/Header";
 export { default as CreateClass } from "./Class/CreateClass/CreateClass";
 export { default as JoinClass } from "./Class/JoinClass/JoinClass";
-// export { default as Login } from "./Class/Login/Login";
+//export { default as LoginTest } from "./Class/Login/LoginTest";
 export { default as JoinedClasses } from "./Class/JoinedClasses/JoinedClasses";
 export { default as Main } from "./Class/Main/Main";
 export { default as Announcment } from "./Class/Announcment/Announcment";
